+++ bin/faucet/src/api/frontend.rs
@@ -13,8 +13,8 @@ pub async fn get_not_found_html() -> Html<&'static str> {
     Html(include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/frontend/not_found.html")))
 }
 
-pub async fn get_index_js() -> JavaScript<&'static str> {
-    JavaScript(include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/frontend/index.js")))
+pub async fn get_bundle_js() -> JavaScript<&'static str> {
+    JavaScript(include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/frontend/bundle.js")))
 }
 
 pub async fn get_index_css() -> Css<&'static str> {
@@ -25,6 +25,14 @@ pub async fn get_background() -> Response {
     (
         [(header::CONTENT_TYPE, header::HeaderValue::from_static("image/png"))],
         include_bytes!(concat!(env!("CARGO_MANIFEST_DIR"), "/frontend/background.png"),),
+    )
+        .into_response()
+}
+
+pub async fn get_wallet_icon() -> Response {
+    (
+        [(header::CONTENT_TYPE, header::HeaderValue::from_static("image/png"))],
+        include_bytes!(concat!(env!("CARGO_MANIFEST_DIR"), "/frontend/wallet-icon.png"),),
     )
         .into_response()
 }